mBox++ (with fixed status communication to cBox)
Ignore-this: 307538455938d142707d3bc77f5f5ded

darcs-hash:20140618114454-594ba-38e94bfe629bd2cea54c5f7f886d7f98c819e982

+++ mBox.cc
@@ -213,8 +213,8 @@ int injectionStartCnt = 0;
 double loopDir  = 1;
 
 struct t_status {
-    double loopPos   ;
-    double errornr   ;
+    unsigned short loopPos   ;
+    unsigned short errornr   ;
 } status ;
 
 struct t_message {
